Remove event_subscriber declaration from event_bus.h
