+++ src/event_bus.cpp
@@ -3,6 +3,7 @@
 #include <algorithm>
 
 #include "debug.h"
+#include "event_subscriber.h"
 
 event_subscriber::~event_subscriber()
 {
