+++ src/event_bus.h
@@ -7,22 +7,7 @@
 
 #include "event.h"
 
-class event_bus;
-
-class event_subscriber
-{
-    public:
-        event_subscriber() = default;
-        event_subscriber( const event_subscriber & ) = delete;
-        event_subscriber &operator=( const event_subscriber & ) = delete;
-        virtual ~event_subscriber();
-        virtual void notify( const cata::event & ) = 0;
-    private:
-        friend class event_bus;
-        void on_subscribe( event_bus * );
-        void on_unsubscribe( event_bus * );
-        event_bus *subscribed_to = nullptr;
-};
+class event_subscriber;
 
 class event_bus
 {
